feat: extract RandomNiceColor helper
